Cleanup, separate out IRLS tests
